feat(Family): Add Source and IsVerified to Family entity and DTOs, refactor mappings
- Added 'Source' and 'IsVerified' properties to Family entity.
- Updated FamilyDto and FamilyDetailDto with new properties.
- Generated and applied new EF Core migration.
- Consolidated FamilyLimitConfiguration mapping into MappingProfile.
- Corrected namespaces and 'using' directives for DTOs in the Families module.
- Removed empty DTOs directory.
