+++ apps/backend/src/Application/Families/Queries/FamilyDto.cs
@@ -1,5 +1,5 @@
 ﻿using backend.Application.Common.Dtos;
-using backend.Application.Families.Dtos;
+using backend.Application.Families.Queries;
 
 namespace backend.Application.Families.Queries;
 
@@ -14,6 +14,8 @@ public class FamilyDto : BaseAuditableDto
     public int? TotalGenerations { get; set; } = null;
     public string? Visibility { get; set; }
     public string? AvatarUrl { get; set; }
+    public string Source { get; set; } = null!;
+    public bool IsVerified { get; set; }
     public List<string> ValidationErrors { get; set; } = [];
     public List<FamilyUserDto> FamilyUsers { get; set; } = [];
     public List<Guid> ManagerIds
